+++ example/dart_helper_utils_example.dart
@@ -1,6 +1,12 @@
 import 'package:dart_helper_utils/dart_helper_utils.dart';
 
 Future<void> main() async {
+  // [1, 2, 3]
+  print(tryToList<int>('[1, 2, 3]'));
+
+  // {hello, world}
+  print(toSet<String>('["hello", "world"]'));
+
   const rawJson = '''
 {
     "name": "John",
@@ -12,7 +18,7 @@ Future<void> main() async {
 ''';
 
   // Example of using dynamic conversions and decode on string.
-  final user = toMap<String, dynamic>(rawJson.decode());
+  final user = toMap<String, dynamic>(rawJson);
 
   // Example of using safe int conversions for dynamic data.
   final walletBalance = toInt(user['wallet']);
@@ -35,6 +41,38 @@ Future<void> main() async {
   print('Status: ${status.code} - ${status.desc}');
   print('Is Success: ${status.isSuccess}');
   print('Is Client Error: ${status.isClientError}');
+
+  const stringToConvert =
+      '123Lorem-Ipsum_is_simply 12DummyText & of THE_PRINTING AND type_setting-industry.';
+
+  // Convert to camelCase
+  print('Convert to Camel Case:');
+  print(stringToConvert.toCamelCase);
+
+  // Convert to pascalCase
+  print('Convert to Pascal Case:');
+  print(stringToConvert.toPascalCase);
+
+  // Convert to snake_case
+  print('Convert to Snake Case:');
+  print(stringToConvert.toSnakeCase);
+
+  // Convert to camel_Snake_Case
+  print('Convert to Camel Snake Case:');
+  print(stringToConvert.toCamelSnakeCase);
+
+  // Convert to kebab-case
+  print('Convert to Kebab Case:');
+  print(stringToConvert.toKebabCase);
+
+  // Convert to Title Case
+  print('Convert to Title Case:');
+  print(stringToConvert.toTitleCase);
+
+  // Unlike the toTitleCase, this one generates a title from string
+  // while ignoring dashes and underscores. Useful for entity name formating.
+  print('Generating a title while ignoring - and _');
+  print(stringToConvert.toTitle);
 
   // Example of using TimeUtils to measure execution duration of
   // a specific task. Works with both sync and async.
@@ -63,7 +101,7 @@ Future<void> main() async {
 
   final result = await TimeUtils.runWithTimeout(
     task: () async {
-      await 5.secDelay;
+      await 1.secDelay;
       return 'Completed';
     },
     timeout: const Duration(seconds: 3),