+++ lib/src/extensions/string_extensions/string_extensions.dart
@@ -1,2 +1,3 @@
+export 'case_conversion.dart';
 export 'general_string_extensions.dart';
 export 'mime_checks.dart';